fix: get autoconfig working
Not sure why it can't see the beans by type or name - perhaps it's an order issue - i.e. tries to create before the datasource beans are created. Checking for class at least makes sure there is a DataSource on the classpath and will then try and wire in the bean.. maybe that's good enough otherwise may need to force the order instead?